Added interface to inputLayer
