+++ src/inputLayer/InputReciever.java
@@ -1,0 +1,7 @@
+package inputLayer;
+
+public interface InputReciever {
+	public int[] getInstruction();
+	void printStartingMessage();
+	void close();
+}
