+++ src/main/Coordinator.java
@@ -1,30 +1,31 @@
 package main;
 import logicLayer.GameLogic;
-import logicLayer.GameMapImpl;
+import logicLayer.GameMap;
 import logicLayer.Position;
-import inputLayer.TextInputReciever;
+import inputLayer.InputReciever;
 import output.TextOutputPrinter;
 
 public class Coordinator {
 	
-	private GameMapImpl gameMap;
-	private TextInputReciever textInput = new TextInputReciever();
+	private GameMap gameMap;
+	private InputReciever input;
 	private TextOutputPrinter textOutput = new TextOutputPrinter();
 	private GameLogic logic;
 	
-	public Coordinator(GameMapImpl gameMap) {
+	public Coordinator(GameMap gameMap, InputReciever inputReceiver) {
 		this.gameMap = gameMap;
+		this.input = inputReceiver;
 		this.logic = new GameLogic(gameMap);
 	}
 	
 	public void startGameLoop() {
-		textInput.printStartingMessage();
+		input.printStartingMessage();
 		for (int i = 0; i<99; i++) {
 			//player turn
 			textOutput.printCurrentTurnOwner(String.valueOf(logic.getCurrentOwner().getType()));
 			while(true) {
 				textOutput.printMap(gameMap);
-				int[] instructions = textInput.getInstruction();
+				int[] instructions = input.getInstruction();
 				performMoveToCommand(instructions);
 				if (checkIfPlayerTurnIsOver() == true){
 					break;
@@ -40,7 +41,7 @@ public class Coordinator {
 			}
 			logic.switchOwner();
 		}
-		textInput.close();
+		input.close();
 	}
 	
 	private void performMoveToCommand(int[] positionArray){
